Move operator and base stuff out of actions.ts
