+++ src/actions/actions.ts
@@ -5,13 +5,10 @@ import { VisualBlockMode } from './../mode/modeVisualBlock';
 import { ModeName } from './../mode/mode';
 import { VisualBlockInsertionType } from './../mode/modeVisualBlock';
 import { Range } from './../motion/range';
-import { TextEditor, EditorScrollByUnit, EditorScrollDirection, CursorMovePosition, CursorMoveByUnit } from './../textEditor';
+import { TextEditor, EditorScrollByUnit, EditorScrollDirection } from './../textEditor';
 import { Register, RegisterMode } from './../register/register';
 import { NumericString } from './../number/numericString';
 import { Position, PositionDiff } from './../motion/position';
-import { PairMatcher } from './../matching/matcher';
-import { QuoteMatcher } from './../matching/quoteMatcher';
-import { TagMatcher } from './../matching/tagMatcher';
 import { Tab, TabCommand } from './../cmd_line/commands/tab';
 import { Configuration } from './../configuration/configuration';
 import { allowVSCodeToPropagateCursorUpdatesAndReturnThem } from '../util';
@@ -21,12 +18,15 @@ import { FileCommand } from './../cmd_line/commands/file';
 import { QuitCommand } from './../cmd_line/commands/quit';
 import * as vscode from 'vscode';
 import * as util from './../util';
+import { RegisterAction } from './base';
+import * as operator from './operator';
+import { BaseAction } from './base';
 
 const is2DArray = function<T>(x: any): x is T[][] {
   return Array.isArray(x[0]);
 };
 
-let compareKeypressSequence = function (one: string[] | string[][], two: string[]): boolean {
+export let compareKeypressSequence = function (one: string[] | string[][], two: string[]): boolean {
   if (is2DArray(one)) {
     for (const sequence of one) {
       if (compareKeypressSequence(sequence, two)) {
@@ -79,63 +79,6 @@ let compareKeypressSequence = function (one: string[] | string[][], two: string[
 
   return true;
 };
-
-export class BaseAction {
-  /**
-   * Can this action be paired with an operator (is it like w in dw)? All
-   * BaseMovements can be, and some more sophisticated commands also can be.
-   */
-  public isMotion = false;
-
-  public canBeRepeatedWithDot = false;
-
-  /**
-   * Modes that this action can be run in.
-   */
-  public modes: ModeName[];
-
-  /**
-   * The sequence of keys you use to trigger the action, or a list of such sequences.
-   */
-  public keys: string[] | string[][];
-
-  public mustBeFirstKey = false;
-
-  /**
-   * The keys pressed at the time that this action was triggered.
-   */
-  public keysPressed: string[] = [];
-
-  /**
-   * Is this action valid in the current Vim state?
-   */
-  public doesActionApply(vimState: VimState, keysPressed: string[]): boolean {
-    if (this.modes.indexOf(vimState.currentMode) === -1) { return false; }
-    if (!compareKeypressSequence(this.keys, keysPressed)) { return false; }
-    if (vimState.recordedState.actionsRun.length > 0 &&
-        this.mustBeFirstKey) { return false; }
-    if (this instanceof BaseOperator && vimState.recordedState.operator) { return false; }
-
-    return true;
-  }
-
-  /**
-   * Could the user be in the process of doing this action.
-   */
-  public couldActionApply(vimState: VimState, keysPressed: string[]): boolean {
-    if (this.modes.indexOf(vimState.currentMode) === -1) { return false; }
-    if (!compareKeypressSequence(this.keys.slice(0, keysPressed.length), keysPressed)) { return false; }
-    if (vimState.recordedState.actionsRun.length > 0 &&
-        this.mustBeFirstKey) { return false; }
-    if (this instanceof BaseOperator && vimState.recordedState.operator) { return false; }
-
-    return true;
-  }
-
-  public toString(): string {
-    return this.keys.join("");
-  }
-}
 
 export class DocumentContentChangeAction extends BaseAction {
   contentChanges: {
@@ -325,77 +268,6 @@ export abstract class BaseCommand extends BaseAction {
 
     return vimState;
   }
-}
-
-export class BaseOperator extends BaseAction {
-    constructor(multicursorIndex?: number) {
-      super();
-      this.multicursorIndex = multicursorIndex;
-    }
-    canBeRepeatedWithDot = true;
-
-    /**
-     * If this is being run in multi cursor mode, the index of the cursor
-     * this operator is being applied to.
-     */
-    multicursorIndex: number | undefined = undefined;
-
-    /**
-     * Run this operator on a range, returning the new location of the cursor.
-     */
-    run(vimState: VimState, start: Position, stop: Position): Promise<VimState> {
-      throw new Error("You need to override this!");
-    }
-}
-
-export enum KeypressState {
-  WaitingOnKeys,
-  NoPossibleMatch
-}
-
-export class Actions {
-
-  /**
-   * Every Vim action will be added here with the @RegisterAction decorator.
-   */
-  public static allActions: { type: typeof BaseAction, action: BaseAction }[] = [];
-
-  /**
-   * Gets the action that should be triggered given a key
-   * sequence.
-   *
-   * If there is a definitive action that matched, returns that action.
-   *
-   * If an action could potentially match if more keys were to be pressed, returns true. (e.g.
-   * you pressed "g" and are about to press "g" action to make the full action "gg".)
-   *
-   * If no action could ever match, returns false.
-   */
-  public static getRelevantAction(keysPressed: string[], vimState: VimState): BaseAction | KeypressState {
-    let couldPotentiallyHaveMatch = false;
-
-    for (const thing of Actions.allActions) {
-      const { type, action } = thing!;
-
-      if (action.doesActionApply(vimState, keysPressed)) {
-        const result = new type();
-
-        result.keysPressed = vimState.recordedState.actionKeys.slice(0);
-
-        return result;
-      }
-
-      if (action.couldActionApply(vimState, keysPressed)) {
-        couldPotentiallyHaveMatch = true;
-      }
-    }
-
-    return couldPotentiallyHaveMatch ? KeypressState.WaitingOnKeys : KeypressState.NoPossibleMatch;
-  }
-}
-
-export function RegisterAction(action: typeof BaseAction): void {
-  Actions.allActions.push({ type: action, action: new action() });
 }
 
 // begin actions
@@ -1259,279 +1131,6 @@ export class CommandSearchBackwards extends BaseCommand {
   }
 }
 
-@RegisterAction
-export class DeleteOperator extends BaseOperator {
-    public keys = ["d"];
-    public modes = [ModeName.Normal, ModeName.Visual, ModeName.VisualLine];
-
-    /**
-     * Deletes from the position of start to 1 past the position of end.
-     */
-    public async delete(start: Position, end: Position, currentMode: ModeName,
-                        registerMode: RegisterMode, vimState: VimState, yank = true): Promise<Position> {
-        if (registerMode === RegisterMode.LineWise) {
-          start = start.getLineBegin();
-          end   = end.getLineEnd();
-        }
-
-        end = new Position(end.line, end.character + 1);
-
-        const isOnLastLine = end.line === TextEditor.getLineCount() - 1;
-
-        // Vim does this weird thing where it allows you to select and delete
-        // the newline character, which it places 1 past the last character
-        // in the line. Here we interpret a character position 1 past the end
-        // as selecting the newline character. Don't allow this in visual block mode
-        if (vimState.currentMode !== ModeName.VisualBlock) {
-          if (end.character === TextEditor.getLineAt(end).text.length + 1) {
-            end = end.getDown(0);
-          }
-        }
-
-        let text = vimState.editor.document.getText(new vscode.Range(start, end));
-
-        // If we delete linewise to the final line of the document, we expect the line
-        // to be removed. This is actually a special case because the newline
-        // character we've selected to delete is the newline on the end of the document,
-        // but we actually delete the newline on the second to last line.
-
-        // Just writing about this is making me more confused. -_-
-
-        // rebornix: johnfn's description about this corner case is perfectly correct. The only catch is
-        // that we definitely don't want to put the EOL in the register. So here we run the `getText`
-        // expression first and then update the start position.
-
-        // Now rebornix is confused as well.
-        if (isOnLastLine &&
-            start.line !== 0 &&
-            registerMode === RegisterMode.LineWise) {
-          start = start.getPreviousLineBegin().getLineEnd();
-        }
-
-        if (registerMode === RegisterMode.LineWise) {
-          // slice final newline in linewise mode - linewise put will add it back.
-          text = text.endsWith("\r\n") ? text.slice(0, -2) : (text.endsWith('\n') ? text.slice(0, -1) : text);
-        }
-
-        if (yank) {
-          Register.put(text, vimState, this.multicursorIndex);
-        }
-
-        let diff = new PositionDiff(0, 0);
-        let resultingPosition: Position;
-
-        if (currentMode === ModeName.Visual) {
-          resultingPosition = Position.EarlierOf(start, end);
-        }
-
-        if (start.character > TextEditor.getLineAt(start).text.length) {
-          resultingPosition = start.getLeft();
-          diff = new PositionDiff(0, -1);
-        } else {
-          resultingPosition = start;
-        }
-
-        if (registerMode === RegisterMode.LineWise) {
-          resultingPosition = resultingPosition.getLineBegin();
-          diff = PositionDiff.NewBOLDiff();
-        }
-
-        vimState.recordedState.transformations.push({
-          type  : "deleteRange",
-          range : new Range(start, end),
-          diff  : diff,
-        });
-
-        return resultingPosition;
-    }
-
-    public async run(vimState: VimState, start: Position, end: Position, yank = true): Promise<VimState> {
-        await this.delete(start, end, vimState.currentMode, vimState.effectiveRegisterMode(), vimState, yank);
-
-        vimState.currentMode = ModeName.Normal;
-
-        /*
-          vimState.cursorPosition      = result;
-          vimState.cursorStartPosition = result;
-        */
-
-        return vimState;
-    }
-}
-
-@RegisterAction
-export class DeleteOperatorVisual extends BaseOperator {
-    public keys = ["D"];
-    public modes = [ModeName.Visual, ModeName.VisualLine];
-
-    public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-      // ensures linewise deletion when in visual mode
-      // see special case in DeleteOperator.delete()
-      vimState.currentRegisterMode = RegisterMode.LineWise;
-
-      return await new DeleteOperator(this.multicursorIndex).run(vimState, start, end);
-    }
-}
-
-@RegisterAction
-export class YankOperator extends BaseOperator {
-    public keys = ["y"];
-    public modes = [ModeName.Normal, ModeName.Visual, ModeName.VisualLine];
-    canBeRepeatedWithDot = false;
-
-    public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-      // Hack to make Surround with y (which takes a motion) work.
-
-      if (vimState.surround) {
-        vimState.surround.range = new Range(start, end);
-        vimState.currentMode = ModeName.SurroundInputMode;
-        vimState.cursorPosition = start;
-        vimState.cursorStartPosition = start;
-
-        return vimState;
-      }
-
-      const originalMode = vimState.currentMode;
-
-      if (start.compareTo(end) <= 0) {
-        end = new Position(end.line, end.character + 1);
-      } else {
-        const tmp = start;
-        start = end;
-        end = tmp;
-
-        end = new Position(end.line, end.character + 1);
-      }
-
-      if (vimState.currentRegisterMode === RegisterMode.LineWise) {
-        start = start.getLineBegin();
-        end = end.getLineEnd();
-      }
-
-      let text = TextEditor.getText(new vscode.Range(start, end));
-
-      // If we selected the newline character, add it as well.
-      if (vimState.currentMode === ModeName.Visual &&
-          end.character === TextEditor.getLineAt(end).text.length + 1) {
-        text = text + "\n";
-      }
-
-      Register.put(text, vimState, this.multicursorIndex);
-
-      vimState.currentMode = ModeName.Normal;
-      vimState.cursorStartPosition = start;
-
-      if (originalMode === ModeName.Normal) {
-        vimState.allCursors = vimState.cursorPositionJustBeforeAnythingHappened.map(x => new Range(x, x));
-      } else {
-        vimState.cursorPosition = start;
-      }
-
-      return vimState;
-    }
-}
-
-@RegisterAction
-export class ShiftYankOperatorVisual extends BaseOperator {
-    public keys = ["Y"];
-    public modes = [ModeName.Visual, ModeName.VisualLine, ModeName.VisualBlock];
-
-    public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-      vimState.currentRegisterMode = RegisterMode.LineWise;
-
-      return await new YankOperator().run(vimState, start, end);
-    }
-}
-
-@RegisterAction
-export class DeleteOperatorXVisual extends BaseOperator {
-  public keys = ["x"];
-  public modes = [ModeName.Visual, ModeName.VisualLine];
-
-  public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-    return await new DeleteOperator(this.multicursorIndex).run(vimState, start, end);
-  }
-}
-
-@RegisterAction
-export class ChangeOperatorSVisual extends BaseOperator {
-  public keys = ["s"];
-  public modes = [ModeName.Visual, ModeName.VisualLine];
-
-  public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-    return await new ChangeOperator().run(vimState, start, end);
-  }
-}
-
-@RegisterAction
-export class FormatOperator extends BaseOperator {
-  public keys = ["="];
-  public modes = [ModeName.Normal, ModeName.Visual, ModeName.VisualLine, ModeName.VisualBlock];
-
-  public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-    vimState.editor.selection = new vscode.Selection(start, end);
-    await vscode.commands.executeCommand("editor.action.formatSelection");
-    let line = vimState.cursorStartPosition.line;
-
-    if (vimState.cursorStartPosition.isAfter(vimState.cursorPosition)) {
-      line = vimState.cursorPosition.line;
-    }
-
-    let newCursorPosition = new Position(line, 0);
-    vimState.cursorPosition = newCursorPosition;
-    vimState.cursorStartPosition = newCursorPosition;
-    vimState.currentMode = ModeName.Normal;
-    return vimState;
-  }
-}
-
-@RegisterAction
-export class UpperCaseOperator extends BaseOperator {
-    public keys = ["U"];
-    public modes = [ModeName.Visual, ModeName.VisualLine];
-
-    public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-      const range = new vscode.Range(start, new Position(end.line, end.character + 1));
-      let text = vimState.editor.document.getText(range);
-
-      await TextEditor.replace(range, text.toUpperCase());
-
-      vimState.currentMode = ModeName.Normal;
-      vimState.cursorPosition = start;
-
-      return vimState;
-    }
-}
-
-@RegisterAction
-export class UpperCaseWithMotion extends UpperCaseOperator {
-  public keys = ["g", "U"];
-  public modes = [ModeName.Normal];
-}
-
-@RegisterAction
-export class LowerCaseOperator extends BaseOperator {
-    public keys = ["u"];
-    public modes = [ModeName.Visual, ModeName.VisualLine];
-
-    public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-      const range = new vscode.Range(start, new Position(end.line, end.character + 1));
-      let text = vimState.editor.document.getText(range);
-
-      await TextEditor.replace(range, text.toLowerCase());
-
-      vimState.currentMode = ModeName.Normal;
-      vimState.cursorPosition = start;
-
-      return vimState;
-    }
-}
-
-@RegisterAction
-export class LowerCaseWithMotion extends LowerCaseOperator {
-  public keys = ["g", "u"];
-  public modes = [ModeName.Normal];
-}
 
 @RegisterAction
 export class MarkCommand extends BaseCommand {
@@ -1545,35 +1144,6 @@ export class MarkCommand extends BaseCommand {
 
     return vimState;
   }
-}
-
-@RegisterAction
-export class ChangeOperator extends BaseOperator {
-    public keys = ["c"];
-    public modes = [ModeName.Normal, ModeName.Visual, ModeName.VisualLine];
-
-    public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-        const isEndOfLine = end.character === end.getLineEnd().character;
-        let state = vimState;
-        // If we delete to EOL, the block cursor would end on the final character,
-        // which means the insert cursor would be one to the left of the end of
-        // the line. We do want to run delete if it is a multiline change though ex. c}
-        if (Position.getLineLength(TextEditor.getLineAt(start).lineNumber) !== 0 || (end.line !== start.line)) {
-          if (isEndOfLine) {
-            state = await new DeleteOperator(this.multicursorIndex).run(vimState, start, end.getLeftThroughLineBreaks());
-          } else {
-            state = await new DeleteOperator(this.multicursorIndex).run(vimState, start, end);
-          }
-        }
-
-        state.currentMode = ModeName.Insert;
-
-        if (isEndOfLine) {
-          state.cursorPosition = state.cursorPosition.getRight();
-        }
-
-        return state;
-    }
 }
 
 @RegisterAction
@@ -1842,93 +1412,13 @@ export class PutCommandVisual extends BaseCommand {
       [start, end] = [end, start];
     }
 
-    const result = await new DeleteOperator(this.multicursorIndex).run(vimState, start, end, false);
+    const result = await new operator.DeleteOperator(this.multicursorIndex).run(vimState, start, end, false);
 
     return await new PutCommand().exec(start, result, true);
   }
 
   // TODO - execWithCount
 }
-
-@RegisterAction
-class IndentOperator extends BaseOperator {
-  modes = [ModeName.Normal];
-  keys = [">"];
-
-  public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-    vimState.editor.selection = new vscode.Selection(start.getLineBegin(), end.getLineEnd());
-
-    await vscode.commands.executeCommand("editor.action.indentLines");
-
-    vimState.currentMode    = ModeName.Normal;
-    vimState.cursorPosition = start.getFirstLineNonBlankChar();
-
-    return vimState;
-  }
-}
-
-/**
- * `3>` to indent a line 3 times in visual mode is actually a bit of a special case.
- *
- * > is an operator, and generally speaking, you don't run operators multiple times, you run motions multiple times.
- * e.g. `d3w` runs `w` 3 times, then runs d once.
- *
- * Same with literally every other operator motion combination... until `3>`in visual mode
- * walked into my life.
- */
-@RegisterAction
-class IndentOperatorInVisualModesIsAWeirdSpecialCase extends BaseOperator {
-  modes = [ModeName.Visual, ModeName.VisualLine];
-  keys = [">"];
-
-  public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-    for (let i = 0; i < (vimState.recordedState.count || 1); i++) {
-      await vscode.commands.executeCommand("editor.action.indentLines");
-    }
-
-    vimState.currentMode    = ModeName.Normal;
-    vimState.cursorPosition = start.getFirstLineNonBlankChar();
-
-    return vimState;
-  }
-}
-
-@RegisterAction
-class OutdentOperator extends BaseOperator {
-  modes = [ModeName.Normal];
-  keys = ["<"];
-
-  public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-    vimState.editor.selection = new vscode.Selection(start, end);
-
-    await vscode.commands.executeCommand("editor.action.outdentLines");
-    vimState.currentMode  = ModeName.Normal;
-    vimState.cursorPosition = start.getFirstLineNonBlankChar();
-
-    return vimState;
-  }
-}
-
-/**
- * See comment for IndentOperatorInVisualModesIsAWeirdSpecialCase
- */
-@RegisterAction
-class OutdentOperatorInVisualModesIsAWeirdSpecialCase extends BaseOperator {
-  modes = [ModeName.Visual, ModeName.VisualLine];
-  keys = ["<"];
-
-  public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-    for (let i = 0; i < (vimState.recordedState.count || 1); i++) {
-      await vscode.commands.executeCommand("editor.action.outdentLines");
-    }
-
-    vimState.currentMode    = ModeName.Normal;
-    vimState.cursorPosition = start.getFirstLineNonBlankChar();
-
-    return vimState;
-  }
-}
-
 
 @RegisterAction
 export class PutBeforeCommand extends BaseCommand {
@@ -2277,7 +1767,7 @@ class CommandDeleteToLineEnd extends BaseCommand {
       return vimState;
     }
 
-    return await new DeleteOperator(this.multicursorIndex).run(vimState, position, position.getLineEnd().getLeft());
+    return await new operator.DeleteOperator(this.multicursorIndex).run(vimState, position, position.getLineEnd().getLeft());
   }
 }
 
@@ -2293,7 +1783,7 @@ export class CommandYankFullLine extends BaseCommand {
 
     vimState.currentRegisterMode = RegisterMode.LineWise;
 
-    return await new YankOperator().run(vimState, start, end);
+    return await new operator.YankOperator().run(vimState, start, end);
   }
 }
 
@@ -2306,7 +1796,7 @@ class CommandChangeToLineEnd extends BaseCommand {
   public async exec(position: Position, vimState: VimState): Promise<VimState> {
     let count = vimState.recordedState.count || 1;
 
-    return new ChangeOperator().run(vimState, position, position.getDownByCount(Math.max(0, count - 1)).getLineEnd().getLeft());
+    return new operator.ChangeOperator().run(vimState, position, position.getDownByCount(Math.max(0, count - 1)).getLineEnd().getLeft());
   }
 }
 
@@ -2319,7 +1809,7 @@ class CommandClearLine extends BaseCommand {
   public async exec(position: Position, vimState: VimState): Promise<VimState> {
     let count = vimState.recordedState.count || 1;
     let end = position.getDownByCount(Math.max(0, count - 1)).getLineEnd().getLeft();
-    return new ChangeOperator().run(vimState, position.getLineBeginRespectingIndent(), end);
+    return new operator.ChangeOperator().run(vimState, position.getLineBeginRespectingIndent(), end);
   }
 }
 
@@ -2757,7 +2247,7 @@ class ActionDeleteChar extends BaseCommand {
       return vimState;
     }
 
-    const state = await new DeleteOperator(this.multicursorIndex).run(vimState, position, position);
+    const state = await new operator.DeleteOperator(this.multicursorIndex).run(vimState, position, position);
 
     state.currentMode = ModeName.Normal;
 
@@ -2778,7 +2268,7 @@ class ActionDeleteCharWithDeleteKey extends BaseCommand {
       return vimState;
     }
 
-    const state = await new DeleteOperator(this.multicursorIndex).run(vimState, position, position);
+    const state = await new operator.DeleteOperator(this.multicursorIndex).run(vimState, position, position);
 
     state.currentMode = ModeName.Normal;
 
@@ -2797,7 +2287,7 @@ class ActionDeleteLastChar extends BaseCommand {
       return vimState;
     }
 
-    return await new DeleteOperator(this.multicursorIndex).run(vimState, position.getLeft(), position.getLeft());
+    return await new operator.DeleteOperator(this.multicursorIndex).run(vimState, position.getLeft(), position.getLeft());
   }
 }
 
@@ -3218,7 +2708,7 @@ class ActionJoinNoWhitespace extends BaseCommand {
 
     let resultLine = lineOne + lineTwo;
 
-    let newState = await new DeleteOperator(this.multicursorIndex).run(
+    let newState = await new operator.DeleteOperator(this.multicursorIndex).run(
       vimState,
       position.getLineBegin(),
       lineTwo.length > 0 ?
@@ -3663,27 +3153,7 @@ class ActionGoToInsertVisualBlockModeAppend extends BaseCommand {
   }
 }
 
-@RegisterAction
-export class YankVisualBlockMode extends BaseOperator {
-    public keys = ["y"];
-    public modes = [ModeName.VisualBlock];
-    canBeRepeatedWithDot = false;
-    runsOnceForEveryCursor() { return false; }
 
-    public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-      let toCopy: string = "";
-
-      for ( const { line } of Position.IterateLine(vimState)) {
-        toCopy += line + '\n';
-      }
-
-      Register.put(toCopy, vimState, this.multicursorIndex);
-
-      vimState.currentMode = ModeName.Normal;
-      vimState.cursorPosition = start;
-      return vimState;
-    }
-}
 
 @RegisterAction
 class InsertInInsertVisualBlockMode extends BaseCommand {
@@ -3745,7 +3215,7 @@ class ActionDeleteLineVisualMode extends BaseCommand {
   keys = ["X"];
 
   public async exec(position: Position, vimState: VimState): Promise<VimState> {
-    return await new DeleteOperator(this.multicursorIndex).run(vimState, position.getLineBegin(), position.getLineEnd());
+    return await new operator.DeleteOperator(this.multicursorIndex).run(vimState, position.getLineBegin(), position.getLineEnd());
   }
 }
 
@@ -3756,7 +3226,7 @@ class ActionChangeChar extends BaseCommand {
   runsOnceForEachCountPrefix = true;
 
   public async exec(position: Position, vimState: VimState): Promise<VimState> {
-    const state = await new ChangeOperator().run(vimState, position, position);
+    const state = await new operator.ChangeOperator().run(vimState, position, position);
 
     state.currentMode = ModeName.Insert;
 
@@ -3776,68 +3246,6 @@ class ActionChangeChar extends BaseCommand {
 }
 
 @RegisterAction
-class ToggleCaseOperator extends BaseOperator {
-  public keys = ["~"];
-  public modes = [ModeName.Visual, ModeName.VisualLine];
-
-  public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-    const range = new vscode.Range(start, end.getRight());
-
-    await ToggleCaseOperator.toggleCase(range);
-
-    const cursorPosition = start.isBefore(end) ? start : end;
-    vimState.cursorPosition = cursorPosition;
-    vimState.cursorStartPosition = cursorPosition;
-    vimState.currentMode = ModeName.Normal;
-
-    return vimState;
-  }
-
-  static async toggleCase(range: vscode.Range) {
-    const text = TextEditor.getText(range);
-
-    let newText = "";
-    for (var i = 0; i < text.length; i++) {
-      var char = text[i];
-      // Try lower-case
-      let toggled = char.toLocaleLowerCase();
-      if (toggled === char) {
-        // Try upper-case
-        toggled = char.toLocaleUpperCase();
-      }
-      newText += toggled;
-    }
-    await TextEditor.replace(range, newText);
-  }
-}
-
-@RegisterAction
-class ToggleCaseVisualBlockOperator extends BaseOperator {
-  public keys = ["~"];
-  public modes = [ModeName.VisualBlock];
-
-  public async run(vimState: VimState, startPos: Position, endPos: Position): Promise<VimState> {
-    for (const { start, end } of Position.IterateLine(vimState)) {
-      const range = new vscode.Range(start, end);
-      await ToggleCaseOperator.toggleCase(range);
-    }
-
-    const cursorPosition = startPos.isBefore(endPos) ? startPos : endPos;
-    vimState.cursorPosition = cursorPosition;
-    vimState.cursorStartPosition = cursorPosition;
-    vimState.currentMode = ModeName.Normal;
-
-    return vimState;
-  }
-}
-
-@RegisterAction
-class ToggleCaseWithMotion extends ToggleCaseOperator {
-  public keys = ["g", "~"];
-  public modes = [ModeName.Normal];
-}
-
-@RegisterAction
 class ToggleCaseAndMoveForward extends BaseCommand {
   modes = [ModeName.Normal];
   keys = ["~"];
@@ -3845,7 +3253,7 @@ class ToggleCaseAndMoveForward extends BaseCommand {
   runsOnceForEachCountPrefix = true;
 
   public async exec(position: Position, vimState: VimState): Promise<VimState> {
-    await new ToggleCaseOperator().run(vimState, vimState.cursorPosition, vimState.cursorPosition);
+    await new operator.ToggleCaseOperator().run(vimState, vimState.cursorPosition, vimState.cursorPosition);
 
     vimState.cursorPosition = vimState.cursorPosition.getRight();
     return vimState;
@@ -4316,38 +3724,4 @@ class MoveEasyMotion extends BaseCommand {
 
     return vimState;
   }
-}
-
-@RegisterAction
-export class CommentOperator extends BaseOperator {
-  public keys = ["g", "b"];
-  public modes = [ModeName.Normal, ModeName.Visual, ModeName.VisualLine];
-
-  public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-    vimState.editor.selection = new vscode.Selection(start.getLineBegin(), end.getLineEnd());
-    await vscode.commands.executeCommand("editor.action.commentLine");
-
-    vimState.cursorPosition = new Position(start.line, 0);
-    vimState.currentMode = ModeName.Normal;
-
-    return vimState;
-  }
-}
-
-@RegisterAction
-export class CommentBlockOperator extends BaseOperator {
-  public keys = ["g", "B"];
-  public modes = [ModeName.Normal, ModeName.Visual, ModeName.VisualLine];
-
-  public async run(vimState: VimState, start: Position, end: Position): Promise<VimState> {
-    const endPosition = vimState.currentMode === ModeName.Normal ? end.getRight() : end;
-    vimState.editor.selection = new vscode.Selection(start, endPosition);
-    await vscode.commands.executeCommand("editor.action.blockComment");
-
-    vimState.cursorPosition = start;
-    vimState.currentMode = ModeName.Normal;
-
-    return vimState;
-  }
-
 }
